search is dotcom: reorganize homepage code
