+++ client/web/src/SourcegraphWebApp.scss
@@ -82,7 +82,7 @@ body,
 @import './global/GlobalDebug';
 @import './global/Notices';
 @import './savedSearches/index';
-@import './search/input/SearchPage';
+@import './search/home/SearchPage';
 @import './search/SearchConsolePage';
 @import './search/ShepherdTour';
 @import './repogroups/RepogroupPage';
