+++ client/web/src/search/home/LoggedOutHomepage.tsx
@@ -12,7 +12,6 @@ import { repogroupList } from '../../repogroups/HomepageConfig'
 
 import { HomepageModalVideo } from './HomepageModalVideo'
 import styles from './LoggedOutHomepage.module.scss'
-import { SignUpCta } from './SignUpCta'
 
 interface SearchExample {
     label: string
@@ -118,84 +117,54 @@ const SearchExamples: React.FunctionComponent<SearchExamplesProps> = ({
     )
 }
 
-export const LoggedOutHomepage: React.FunctionComponent<LoggedOutHomepageProps> = props => {
-    const onClickInstallSubtext = useCallback(() => {
-        props.telemetryService.log(
-            'HomepageInstallSourcegraphCTAClicked',
-            { name: 'InstallSourcegraphSubtext' },
-            { name: 'InstallSourcegraphSubtext' }
-        )
-    }, [props.telemetryService])
-
-    return (
-        <div className={styles.loggedOutHomepage}>
-            <div className={styles.helpContent}>
-                {props.featureFlags.get('search-notebook-onboarding') ? (
-                    <SearchExamples
-                        title="Search notebooks"
-                        subtitle="Three ways code search is more efficient than your IDE"
-                        examples={exampleNotebooks}
-                        icon={<BookOutlineIcon />}
-                        {...props}
-                    />
-                ) : (
-                    <SearchExamples
-                        title="Search examples"
-                        subtitle="Find answers faster with code search across multiple repos and commits"
-                        examples={exampleQueries}
-                        icon={<MagnifyingGlassSearchIcon />}
-                        {...props}
-                    />
-                )}
-                <div className={styles.thumbnail}>
-                    <div className={classNames(styles.title, 'mb-2')}>Watch and learn</div>
-                    <HomepageModalVideo {...props} />
-                </div>
-            </div>
-
-            <div className="mt-5 d-flex justify-content-center">
-                <div className="d-flex align-items-center flex-column">
-                    <SignUpCta className={styles.loggedOutHomepageCta} telemetryService={props.telemetryService} />
-                    <div className="mt-2 text-center">
-                        Search private code by{' '}
-                        <a
-                            href="https://docs.sourcegraph.com/admin/install"
-                            target="_blank"
-                            rel="noopener noreferrer"
-                            onClick={onClickInstallSubtext}
-                        >
-                            installing Sourcegraph locally.
-                        </a>
-                    </div>
-                </div>
-            </div>
-
-            <div className="mt-5">
-                <div className="d-flex align-items-baseline mt-5 mb-3">
-                    <div className={classNames(styles.title, 'mr-2')}>Repository groups</div>
-                    <div className="font-weight-normal text-muted">Search sets of repositories</div>
-                </div>
-                <div className={styles.loggedOutHomepageRepogroupListCards}>
-                    {repogroupList.map(repogroup => (
-                        <div className="d-flex align-items-center" key={repogroup.name}>
-                            <img
-                                className={classNames(styles.loggedOutHomepageRepogroupListIcon, 'mr-2')}
-                                src={repogroup.homepageIcon}
-                                alt={`${repogroup.name} icon`}
-                            />
-                            <Link
-                                to={repogroup.url}
-                                className={classNames(styles.loggedOutHomepageRepogroupListingTitle)}
-                            >
-                                {repogroup.title}
-                            </Link>
-                        </div>
-                    ))}
-                </div>
+export const LoggedOutHomepage: React.FunctionComponent<LoggedOutHomepageProps> = props => (
+    <div className={styles.loggedOutHomepage}>
+        <div className={styles.helpContent}>
+            {props.featureFlags.get('search-notebook-onboarding') ? (
+                <SearchExamples
+                    title="Search notebooks"
+                    subtitle="Three ways code search is more efficient than your IDE"
+                    examples={exampleNotebooks}
+                    icon={<BookOutlineIcon />}
+                    {...props}
+                />
+            ) : (
+                <SearchExamples
+                    title="Search examples"
+                    subtitle="Find answers faster with code search across multiple repos and commits"
+                    examples={exampleQueries}
+                    icon={<MagnifyingGlassSearchIcon />}
+                    {...props}
+                />
+            )}
+            <div className={styles.thumbnail}>
+                <div className={classNames(styles.title, 'mb-2')}>Watch and learn</div>
+                <HomepageModalVideo {...props} />
             </div>
         </div>
-    )
-}
+
+        <div className="mt-5">
+            <div className="d-flex align-items-baseline mt-5 mb-3">
+                <div className={classNames(styles.title, 'mr-2')}>Repository groups</div>
+                <div className="font-weight-normal text-muted">Search sets of repositories</div>
+            </div>
+            <div className={styles.loggedOutHomepageRepogroupListCards}>
+                {repogroupList.map(repogroup => (
+                    <div className="d-flex align-items-center" key={repogroup.name}>
+                        <img
+                            className={classNames(styles.loggedOutHomepageRepogroupListIcon, 'mr-2')}
+                            src={repogroup.homepageIcon}
+                            alt={`${repogroup.name} icon`}
+                        />
+                        <Link to={repogroup.url} className={classNames(styles.loggedOutHomepageRepogroupListingTitle)}>
+                            {repogroup.title}
+                        </Link>
+                    </div>
+                ))}
+            </div>
+        </div>
+    </div>
+)
 
 const MagnifyingGlassSearchIcon = React.memo(() => (
     <svg width="18" height="18" fill="none" xmlns="http://www.w3.org/2000/svg">
