+++ client/web/src/search/home/SearchPageFooter.test.tsx
