+++ client/web/src/search/home/SearchPageFooter.tsx
