+++ client/web/src/search/home/homepage.scss
